// src/types.ts
export interface LiturgyReading {
  ref: string; 
  text: string; 
}

export interface LiturgyDay {
  date: string;
  liturgicalColor: string;
  season: string;
  saint: string; 
  readings: {
    first: LiturgyReading;
    psalm: LiturgyReading; 
    second?: LiturgyReading; 
    gospel: LiturgyReading;
  };
}